+++ crates/common/src/timer.rs
@@ -126,11 +126,20 @@ impl PartialEq for TimeEvent {
 
 pub type RustTimeEventCallback = dyn Fn(TimeEvent);
 
-#[derive(Clone)]
 pub enum TimeEventCallback {
     #[cfg(feature = "python")]
-    Python(Arc<PyObject>),
+    Python(PyObject),
     Rust(Rc<RustTimeEventCallback>),
+}
+
+impl Clone for TimeEventCallback {
+    fn clone(&self) -> Self {
+        match self {
+            #[cfg(feature = "python")]
+            Self::Python(obj) => Self::Python(nautilus_core::python::clone_py_object(obj)),
+            Self::Rust(cb) => Self::Rust(cb.clone()),
+        }
+    }
 }
 
 impl Debug for TimeEventCallback {
@@ -171,7 +180,7 @@ impl From<Rc<RustTimeEventCallback>> for TimeEventCallback {
 #[cfg(feature = "python")]
 impl From<PyObject> for TimeEventCallback {
     fn from(value: PyObject) -> Self {
-        Self::Python(Arc::new(value))
+        Self::Python(value)
     }
 }